# LaravelProject2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/createNote": {
            "post": {
                "summary": "Create Note",
                "description": " Create Note ",
                "operationId": "3f1be91984161a01ea4647576ef3b3d6",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "notes created successfully"
                    },
                    "400": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/getNoteById": {
            "post": {
                "summary": "Read Note",
                "description": "Read Notes For an Particular User",
                "operationId": "113d4a9232db6d7251c0eb0df352bbd5",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "All Notes are found Successfully"
                    },
                    "404": {
                        "description": "Notes Not Found"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/getAllNotes": {
            "get": {
                "summary": "get All Notes",
                "description": "get All Notes",
                "operationId": "e245bcac922abd123d138953b1261495",
                "responses": {
                    "200": {
                        "description": "Notes found successfully"
                    },
                    "404": {
                        "description": "Invalid Autherization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updateNoteById": {
            "post": {
                "summary": "update note",
                "description": "update note",
                "operationId": "a11e25c459bae96f3c862f767bfb8430",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note Updated Successfully"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    },
                    "404": {
                        "description": "Notes Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deleteNoteById": {
            "delete": {
                "summary": "delete note",
                "description": "delete user note",
                "operationId": "563ccd06fc5b23af9eea59ab50ff51fc",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note successfully deleted"
                    },
                    "404": {
                        "description": "Notes not found"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/addNoteLabel": {
            "post": {
                "summary": "Add note label",
                "description": "Adiing note label",
                "operationId": "38fa105a82f7c3e3eeac4714f589438c",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "label_id",
                                    "note_id"
                                ],
                                "properties": {
                                    "label_id": {
                                        "type": "string"
                                    },
                                    "note_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note successfully updated"
                    },
                    "402": {
                        "description": "Labels or Notes not found"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/searchNotes": {
            "post": {
                "summary": "Search Note",
                "description": " Search Note ",
                "operationId": "0af3875fb0e9e0a7865d06f95d9dc2e7",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "search"
                                ],
                                "properties": {
                                    "search": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note Fetched Sucessfully"
                    },
                    "404": {
                        "description": "Notes Not Found"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "laravel bearer authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}